requirements.txt
pypdf
openai
langchain
llama_index
python-dotenv
python-decouple
azure-core
tiktoken
fastapi
uvicorn
nltk